+++ src/App.js
@@ -1,12 +1,10 @@
 import MenuDone from './Components/MenuDone/MenuDome';
-import Router from './Router/Router'
-
-
+import Router from './Router/Router';
 function App() {
   return (
     <div className="App">
       <MenuDone />
-      <Router/>
+      <Router />
     </div>
   );
 }
