+++ src/Utils/utils.scss
@@ -16,7 +16,8 @@ $white-color: #ffffff;
 $button-color: #ff851c;
 $cancel-button: #24d40c;
 $delete-button: #db0938;
-
+$logout-color: #3e4e6c;
+$header-text: #7b8aa4;
 //Цвета категорий
 $family-color: #ffe6d3;
 $healt-color: #cdf7ff;
@@ -82,4 +83,12 @@ $active-color: #b9c3c8;
   @media screen and (min-width: 1280px) {
     grid-template-columns: repeat(5, 1fr);
   }
+}
+
+%img-styles {
+  width: 100%;
+  height: 100%;
+  display: block;
+  object-fit: cover;
+  object-position: center;
 }
